# Disable `bundleStrings` RPC extension