@msujew 2022-09-05
cc @tortmayr I'm open for any other ideas on how to address this issue. Feel free to suggest something else than disabling the `bundleStrings` extension.