@tortmayr 2022-09-06
Change looks good to me, I can confirm that plugin support now works as expected after switching the language.

>cc @tortmayr I'm open for any other ideas on how to address this issue. Feel free to suggest something else than disabling the bundleStrings extension.

I also looked into this issue and for the moment it looks like the only option is to disable the `bundleStrings` extension. Luckily for us the performance hit should is only  minor because only the browser string decoding/encoding is affected. In the backend we can still rely on the fast node string parsing.

I also extended and rerun the benchmark extension (introduced for https://github.com/eclipse-theia/theia/issues/11159):

# Results

Testruns: 500
| TestCase | 1.26 (JsonRpc) | 1.28 (custom codec) | Master(Msgpackr) | PR (Msgpackr-BundleStrings) | Description |
|----------|-----------------------|-----------------------|---------------------|-----|-----|
| Small | 0.80 ms<br> (SD: 3.18) | 1.00 ms<br> (SD: 8.01) | 0.90 ms<br> (SD: 2.3) | 0.9 ms<br> (SD: 1.95) |Simple object with 3 properties of primitive type|
| Medium | 0.90 ms<br> (SD: 3.52) | 1.15 ms<br> (SD: 6.78) | 1.00 ms<br> (SD: 2.52) |1.0 ms<br> (SD: 2.68)|Array that contains 6 objects with nested properties |
| Large | 148.60 ms<br> (SD: 11.56) | 345.85 ms<br> (SD: 49.52) | 42.60 ms<br> (SD: 8.87) | 69.55 ms<br> (SD: 13.61)|Large Object (8MB serialized). (Captured result from a `getDeployedPlugins` request in the example application)|
| Buffer | 850.6 ms<br> (SD: 27.05) | 17.30 ms<br> (SD: 23.52) | 16.00 ms<br> (SD: 8.18) |15.80 ms<br> (SD: 7.22) |8MB binary buffer
| LargeString| 57.80 ms<br> (SD: 7.94) |226.65 ms<br> (SD: 42.61) | 15.20 ms<br> (SD: 6.65)| 20.10 ms<br> (SD: 7.19)|Large object with lots of string data (4MB serialized)|
| RandomString|29.75 ms<br> (SD: 6.88) | 7.55 ms<br> (SD: 13.86)| 6.70 ms<br> (SD: 5.09)| 6.70 ms<br> (SD: 5.19)|Array containing 100 random strings of length 1000|

It looks like disabling the `bundleStrings` option has almost no performance impact on small to medium objects. The performance degradation is only really noticeable in really large objects with lots of really long strings. And even then the performance is still way better than what we would achieve with the old JsonRpc or custom message codec.

I'm going to open an issue for this in the `msgpackr` repository. If a fix for this becomes available we can reactivate the `bundleStrings` options.