@msujew 2022-09-06
@tortmayr Thanks for looking into it and rerunning the benchmarks! Really appreciated :)